Arreglos y float

+++ Entregable 2/Src/LibraryModelado/Factory/FactoryCSharp/FactoryButtonGeneric.cs
@@ -35,22 +35,22 @@ namespace Proyecto.Factory.CSharp
         /// <summary>
         /// Ancho del boton.
         /// </summary>
-        private int width;
+        private float width;
 
         /// <summary>
         /// Altura del boton.
         /// </summary>
-        private int height;
+        private float height;
 
         /// <summary>
         /// Posicion en X del boton.
         /// </summary>
-        private int positionX;
+        private float positionX;
 
         /// <summary>
         /// Posicion en Y del boton.
         /// </summary>
-        private int positionY;
+        private float positionY;
 
         /// <summary>
         /// Nivel del boton.
@@ -71,10 +71,10 @@ namespace Proyecto.Factory.CSharp
         public override IComponent MakeComponent(Tag tag)
         {
             this.name = tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Name"; }).Valor;
-            this.width = Convert.ToInt32(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Width"; }).Valor);
-            this.height = Convert.ToInt32(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Height"; }).Valor);
-            this.positionX = Convert.ToInt32(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "PositionX"; }).Valor);
-            this.positionY = Convert.ToInt32(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "PositionY"; }).Valor);
+            this.width = Convert.ToSingle(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Width"; }).Valor);
+            this.height = Convert.ToSingle(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Height"; }).Valor);
+            this.positionX = Convert.ToSingle(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "PositionX"; }).Valor);
+            this.positionY = Convert.ToSingle(tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "PositionY"; }).Valor);
             this.color = tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Color"; }).Valor;
             this.image = tag.Atributos.Find(delegate(Atributos atr) { return atr.Clave == "Photo"; }).Valor;
             this.level = this.world.SpaceList.Last();
