+++ Entregable 2/Src/LibraryModelado/Factory/FactoryCSharp/FactoryMoneyContainer.cs
@@ -31,22 +31,22 @@ namespace Proyecto.Factory.CSharp
         /// <summary>
         /// Ancho del container.
         /// </summary>
-        private int width;
+        private float width;
 
         /// <summary>
         /// Altura del container.
         /// </summary>
-        private int height;
+        private float height;
 
         /// <summary>
         /// Posicion en X del container.
         /// </summary>
-        private int positionX;
+        private float positionX;
 
         /// <summary>
         /// Posicion en Y del container.
         /// </summary>
-        private int positionY;
+        private float positionY;
 
         /// <summary>
         /// Nivel al que pertenece el item.
@@ -56,7 +56,7 @@ namespace Proyecto.Factory.CSharp
         /// <summary>
         /// Valor que aceptara el container.
         /// </summary>
-        private int acceptableValue = 0;
+        private float acceptableValue = 0;
 
         /// <summary>
         /// Instancia del mundo.
@@ -71,14 +71,15 @@ namespace Proyecto.Factory.CSharp
         /// <returns>Componente <see cref="IComponent"/>.</returns>
         public override IComponent MakeComponent(Tag tag)
         {
+            
             this.name = tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Name"; }).Valor;
             this.level = this.world.SpaceList.Last();
-            this.width = Convert.ToInt32(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Width"; }).Valor);
-            this.height = Convert.ToInt32(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Height"; }).Valor);
-            this.positionX = Convert.ToInt32(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "PositionX"; }).Valor);
-            this.positionY = Convert.ToInt32(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "PositionY"; }).Valor);
+            this.width = Convert.ToSingle(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Width"; }).Valor);
+            this.height = Convert.ToSingle(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Height"; }).Valor);
+            this.positionX = Convert.ToSingle(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "PositionX"; }).Valor);
+            this.positionY = Convert.ToSingle(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "PositionY"; }).Valor);
             this.image = tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Photo"; }).Valor;
-            this.acceptableValue = Convert.ToInt32(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Value"; }).Valor);
+            this.acceptableValue = Convert.ToSingle(tag.Atributos.Find(delegate (Atributos atr) { return atr.Clave == "Value"; }).Valor);
             Items moneyContainer = new MoneyContainer(this.name, this.level, this.positionX, this.positionY, this.width, this.height, this.image, this.acceptableValue);
             this.level.ItemList.Add(moneyContainer);
             return moneyContainer;
